Use Java 8 standards, up-to-date Spring standards, and AssertJ instea...

+++ src/main/java/com/exo/ecommerce/domain/invoice/Invoice.java
@@ -36,8 +36,7 @@ public class Invoice {
         return cart;
     }
 
-    public float calculateTotalPrice()
-    {
+    public double calculateTotalPrice() {
         return this.cart.calculateTotalPrice();
     }
